Split mod_p2/mod_p3/mod_p5 at clean boundaries to clear max-lines
